requirements.txt
allure-pytest
pytest
requests
selenium
urllib3